BigDecimal para variables de moneda

+++ src/main/java/com/josefuentes4096/products/dto/OrderResponseDTO.java
@@ -4,6 +4,7 @@ import com.josefuentes4096.products.enums.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,6 +13,6 @@ public class OrderResponseDTO {
     private Integer orderId;
     private Integer userId;
     private OrderStatus status;
-    private Double total;
+    private BigDecimal total;
     private List<OrderItemResponseDTO> items;
 }
